+++ CHANGES.rst
@@ -45,6 +45,8 @@ astropy.io.votable
 astropy.modeling
 ^^^^^^^^^^^^^^^^
 
+- Fixed use of weights with ``LinearLSQFitter``. [#10687]
+
 astropy.nddata
 ^^^^^^^^^^^^^^